services/cross-chain-integration/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
google-cloud-pubsub==2.25.0
alembic==1.13.2
web3==7.14.0
eth-account==0.13.7
eth-abi>=5.0.1
